+++ src/main/java/com/lucy/controller/CreditAccountController.java
@@ -72,7 +72,7 @@ public class SavingAccountController {
 	}
 
 	@RequestMapping(value = "/deleteAccount/{id}", method = RequestMethod.DELETE)
-	public String deleteAccount(@ModelAttribute("account") CreditAccount account, @RequestParam() Long id, Model model,
+	public String deleteAccount(@ModelAttribute("account") CreditAccount account, @RequestParam() Integer id, Model model,
 			RedirectAttributes redirectAttributes) {
 		creditAccountService.delete(id);
 		return "redirect:/credit/success";
